cut inv
Use ;:inv as cut's monadic obverse

It would be nice if we could define this obverse to be sensitive to cut's optional left argument, but I currently do not see how to accomplish that.  So... leave obverse for the dyadic case as a domain error to allow for possible future enhancement.